movie-prediction-pipeline/requirements.txt
# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
statsmodels==0.14.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# AWS and SageMaker
boto3==1.28.25
sagemaker==2.175.0
sagemaker-experiments==0.1.45

# Configuration and utilities
PyYAML==6.0.1
joblib==1.3.2
argparse

# Logging and monitoring
mlflow==2.5.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Data validation
great-expectations==0.17.12
